requirements.txt
albumentations==1.1.0
docx==0.2.4
ITMO_FS==0.3.3
joblib==1.1.0
matplotlib==3.5.1
mlflow==1.25.1
nibabel==3.2.2
numpy==1.17.4
opencv_python_headless==4.5.5.64
pandas==1.4.2
Pillow==9.1.1
progressbar33==2.4
pydicom==2.3.0
pymrmre==1.0.7
pyradiomics==3.0.1
python_docx==0.8.11
pytorch_lightning==1.6.3
radiomics==0.1
rich==12.4.4
rpy2==3.5.1
scikit_image==0.19.2
scikit_learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
SimpleITK==2.1.1.2
skimage==0.0
termcolor==1.1.0
torch==1.11.0+cu113
torchsummary==1.5.1
torchvision==0.12.0+cu113
